API: do not call `matplotlib.pyplot.show` inside `plot_partition`
because it is implicit. The function `matplotlib.pyplot.show`
is to be called in user code, not within a package.
Moreover, this call raises a warning when the backend of
`matplotlib` is set to `'Agg'`, which is what happens in
one of `tulip`'s tests:

```
abstract_test.py::transition_directions_test
  .../lib/python3.9/site-packages/polytope/plot.py:133:
UserWarning: Matplotlib is currently using agg, which is a non-GUI
backend, so cannot show the figure.
    mpl.pyplot.show()
```